Emojify returns original cost instead of null
If emojify fails to find the emote and would return null, it will instead return the original cost that was passed to the function.
P.S. I feel like this is a bad implementation of this fix, but I wrote this in 5 mins and I don't think it really matters in the greater scheme of things. Also, it's untested!